세션 달기, 로그인 권한

+++ src/main/java/com/example/modelboard/controller/BoardController.java
@@ -4,14 +4,13 @@ import com.example.modelboard.dto.BoardDTO;
 import com.example.modelboard.service.BoardService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import java.util.Date;
 
 @Controller
+@SessionAttributes("user")
 public class BoardController {
 
     final BoardService boardService;
@@ -28,7 +27,13 @@ public class BoardController {
     }
 
     @GetMapping("/write")
-    public String write() {return "board/write";}
+    public String write(HttpSession session) {
+        if (session.getAttribute("user") != null) {
+            return "board/write";
+        } else {
+            return "redirect:/login";
+        }
+    }
 
     @PostMapping("/write")
     public String writeComplete(BoardDTO boardDTO) {
@@ -39,36 +44,48 @@ public class BoardController {
     }
 
     @GetMapping("/read")
-    public String read(Long bno, Model model){
-        BoardDTO boardDTO = boardService.read(bno);
-        model.addAttribute("board",boardDTO);
-        return "/board/read";
+    public String read(Long bno, Model model, HttpSession session) {
+        if (session.getAttribute("user") != null) {
+            BoardDTO boardDTO = boardService.read(bno);
+            model.addAttribute("board", boardDTO);
+            return "/board/read";
+        } else {
+            return "redirect:/login";
+        }
     }
 
     @GetMapping("/updateBoard")
-    public String updateBoard(Long bno,Model model){
-        BoardDTO boardDTO = boardService.read(bno);
-        model.addAttribute("board",boardDTO);
-        return "/board/boardUpdate";
+    public String updateBoard(Long bno, Model model, HttpSession session) {
+        if (session.getAttribute("user") != null) {
+            BoardDTO boardDTO = boardService.read(bno);
+            model.addAttribute("board", boardDTO);
+            return "/board/boardUpdate";
+        } else {
+            return "redirect:/login";
+        }
     }
 
 
     @PostMapping("/updateBoard")
     @ResponseBody
-    public String updateBoard(@RequestParam("bno") Long bno, @RequestParam("title") String title, @RequestParam("content") String content){
-       BoardDTO boardDTO = boardService.read(bno);
-       boardDTO.setTitle(title);
-       boardDTO.setContent(content);
-       boardDTO.setRegdate(new Date());
-       boardService.updateBoard(boardDTO);
-       return "/read?bno="+bno;
+    public String updateBoard(@RequestParam("bno") Long bno, @RequestParam("title") String title, @RequestParam("content") String content) {
+        BoardDTO boardDTO = boardService.read(bno);
+        boardDTO.setTitle(title);
+        boardDTO.setContent(content);
+        boardDTO.setRegdate(new Date());
+        boardService.updateBoard(boardDTO);
+        return "/read?bno=" + bno;
     }
 
     @PostMapping("/deleteBoard")
     @ResponseBody
-    public String deleteBoard(@RequestParam("bno")Long bno){
-        boardService.deleteBoard(bno);
-        return "/boardList";
+    public String deleteBoard(@RequestParam("bno") Long bno, HttpSession session) {
+        if (session.getAttribute("user") != null) {
+            boardService.deleteBoard(bno);
+            return "/boardList";
+        } else {
+            return "redirect:/login";
         }
     }
+}
